requirements.txt
bcrypt==3.2.2
Django==4.0.5
django-filter==21.1
djangorestframework==3.13.1
psycopg2-binary==2.8.6
python-dotenv==0.20.0
sqlparse==0.4.2

